Normalização dos WS de Pessoas e Lotacoes
